# docs/swagger.yaml
basePath: /api
definitions:
  gorm.DeletedAt:
    properties:
      time:
        type: string
      valid:
        description: Valid is true if Time is not NULL
        type: boolean
    type: object
...
  handlers.CreatePodcastRequest:
    properties:
      category:
        type: string
      description:
        type: string
      duration:
        description: 'Format: MM:SS atau HH:MM:SS'
        type: string
      episode_number:
        type: integer
      host:
        type: string
      release_date:
        description: 'Format: "2006-01-02"'
        type: string
      season:
        type: integer
      thumbnail:
        type: string
      title:
        type: string
      video_url:
        type: string
    required:
    - category
    - description
    - duration
    - host
    - release_date
    - title
    - video_url
    type: object
...
  handlers.UpdatePodcastRequest:
    properties:
      category:
        type: string
      description:
        type: string
      duration:
        description: 'Format: MM:SS atau HH:MM:SS'
        type: string
      episode_number:
        type: integer
      host:
        type: string
      release_date:
        description: 'Format: "2006-01-02"'
        type: string
      season:
        type: integer
      thumbnail:
        type: string
      title:
        type: string
      video_url:
        type: string
    type: object
...
  models.Podcast:
    properties:
      category:
        type: string
      created_at:
        type: string
      deleted_at:
        $ref: '#/definitions/gorm.DeletedAt'
      description:
        type: string
      duration:
        description: 'Format: MM:SS atau HH:MM:SS'
        type: string
      episode_number:
        type: integer
      host:
        type: string
      id:
        type: integer
      release_date:
        type: string
      season:
        type: integer
      thumbnail:
        type: string
      title:
        type: string
      updated_at:
        type: string
      video_url:
        type: string
    type: object
info:
  contact:
    email: support@soundcave.com
    name: API Support
  description: API documentation for SoundCave Backend - Music streaming platform
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: SoundCave Backend API
  version: "1.0"
paths:
...
  /podcasts:
    get:
      consumes:
      - application/json
      description: Get paginated list of podcasts with filtering and search
      parameters:
      - default: 1
        description: Page number
        in: query
        name: page
        type: integer
      - default: 10
        description: Items per page
        in: query
        name: limit
        type: integer
      - description: Filter by host
        in: query
        name: host
        type: string
      - description: Filter by category
        in: query
        name: category
        type: string
      - description: Filter by season
        in: query
        name: season
        type: integer
      - description: Search by title, host, or description
        in: query
        name: search
        type: string
      - default: created_at
        description: Sort field
        in: query
        name: sort_by
        type: string
      - default: desc
        description: Sort order
        in: query
        name: order
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Returns a list of podcasts with pagination
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Get all podcasts
      tags:
      - Podcasts
    post:
      consumes:
      - application/json
      description: Create a new podcast episode
      parameters:
      - description: Podcast Request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/handlers.CreatePodcastRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Create new podcast
      tags:
      - Podcasts
  /podcasts/{id}:
    delete:
      consumes:
      - application/json
      description: Soft delete a podcast by ID
      parameters:
      - description: Podcast ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Delete podcast
      tags:
      - Podcasts
    get:
      consumes:
      - application/json
      description: Get podcast details by ID
      parameters:
      - description: Podcast ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Podcast'
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Get podcast by ID
      tags:
      - Podcasts
    put:
      consumes:
      - application/json
      description: Update podcast information
      parameters:
      - description: Podcast ID
        in: path
        name: id
        required: true
        type: integer
      - description: Update Podcast Request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/handlers.UpdatePodcastRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Podcast'
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Update podcast
      tags:
      - Podcasts
...
securityDefinitions:
  BearerAuth:
    description: 'To authenticate, provide the API token with the ''Bearer '' prefix.
      Example: ''Bearer your_token_here'''
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"